+++ tests/estimation/test_tie.py
@@ -243,6 +243,23 @@ class TestTieAlgorithm(unittest.TestCase):
         with self.assertRaises(ValueError):
             tie.estimateZk(intra)
 
+    def testDefocalOffsetPropagation(self):
+        """Test that image-level defocal offsets propagate.
+
+        We will do this by setting defocal offsets to 0 and making
+        sure an error is raised.
+        """
+        zkTrue, intra, extra = forwardModelPair()
+        intra.defocalOffset = 0
+        extra.defocalOffset = 0
+
+        tie = TieAlgorithm()
+        try:
+            tie.estimateZk(intra, extra)
+            raise RuntimeError("This should have raised an error!")
+        except Exception:
+            pass
+
 
 if __name__ == "__main__":
     # Do the unit test